+++ js/db/users.js
@@ -1,0 +1,5 @@
+export const user = {
+  email: "test@codeit.com",
+  password: "codeit101"
+}
+
